Project2/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
matplotlib
seaborn
networkx
requests
python-dotenv
beautifulsoup4
lxml
pillow
langchain
langchain-core
langchain-google-genai
google-generativeai
openpyxl
pyarrow
tabulate
html5lib
duckdb
psutil
httpx
scikit-learn
pydantic
langchain-openai
langchain-community
langsmith
scipy
openai
tiktoken
faiss-cpu
chromadb
jinja2
